Push the class to package level
