+++ server/src/main/java/io/spine/server/entity/SilentWitness.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2019, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.entity;
+
+import io.spine.validate.NonValidated;
+
+/**
+ * An implementation of a {@code TransactionListener} which does not set any behavior for its
+ * callbacks.
+ */
+public final class SilentWitness<I> implements TransactionListener<I> {
+
+    @Override
+    public void onBeforePhase(Phase<I, ?> phase) {
+        // Do nothing.
+    }
+
+    @Override
+    public void onAfterPhase(Phase<I, ?> phase) {
+        // Do nothing.
+    }
+
+    @Override
+    public void onBeforeCommit(@NonValidated EntityRecord entityRecord) {
+        // Do nothing.
+    }
+
+    @Override
+    public void onTransactionFailed(Throwable t,
+                                    EntityRecord record) {
+        // Do nothing.
+    }
+
+    @Override
+    public void onAfterCommit(EntityRecordChange change) {
+        // Do nothing.
+    }
+}
